requirements.txt
openai
requests
mysql-connector-python
flask
